Updated security policy, fixed attachments.
