+++ src/main/java/business/controllers/LabController.java
@@ -6,16 +6,19 @@ import java.util.List;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
+import business.exceptions.InvalidLabNumber;
+import business.exceptions.LabNotFound;
+import business.exceptions.LabuserWithoutLab;
+import business.exceptions.UserUnauthorised;
 import business.models.ContactData;
 import business.models.Lab;
 import business.models.LabRepository;
@@ -49,13 +52,6 @@ public class LabController {
         lab.getContactData().copy(body.getContactData());
     }
    
-    @ResponseStatus(value=HttpStatus.BAD_REQUEST)  // 
-    public class InvalidLabNumber extends RuntimeException {
-        public InvalidLabNumber() {
-            super("Invalid lab number.");
-        }
-    }
-    
     @RequestMapping(value = "/admin/labs", method = RequestMethod.POST)
     public Lab create(Principal principal, @RequestBody Lab body) {
         log.info("POST /admin/labs (for user: " + principal.getName() + ")");
@@ -67,14 +63,6 @@ public class LabController {
         transferLabData(body, lab);
         return labRepository.save(lab);
     }
-    
-    @ResponseStatus(value=HttpStatus.NOT_FOUND)  // 
-    public class LabNotFound extends RuntimeException {
-        public LabNotFound() {
-            super("Lab cannot be found.");
-        }
-    }
-
     
     @RequestMapping(value = "/admin/labs/{id}", method = RequestMethod.PUT,
             consumes = MediaType.APPLICATION_JSON_VALUE)
@@ -113,20 +101,6 @@ public class LabController {
         return  labRepository.save(lab);
     }
     
-    @ResponseStatus(value=HttpStatus.UNAUTHORIZED)  // 
-    public class UserUnauthorised extends RuntimeException {
-        public UserUnauthorised(String message) {
-            super(message);
-        }
-    }
-    
-    @ResponseStatus(value=HttpStatus.BAD_REQUEST)  // 
-    public class LabuserWithoutLab extends RuntimeException {
-        public LabuserWithoutLab(String message) {
-            super(message);
-        }
-    }
-    
     @RequestMapping(value = "/lab", method = RequestMethod.GET)
     public Lab getLab(UserAuthenticationToken token) {
         User user = token.getUser();
@@ -139,6 +113,7 @@ public class LabController {
         throw new UserUnauthorised("User not authorised to fetch lab information.");
     }
     
+    @PreAuthorize("isAuthenticated() and hasPermission('isLabuser')")
     @RequestMapping(value = "/lab", method = RequestMethod.PUT,
             consumes = MediaType.APPLICATION_JSON_VALUE)
     public Lab updateLab(UserAuthenticationToken token, @RequestBody Lab body) {
