+++ src/main/java/business/exceptions/InvalidLabNumber.java
@@ -1,0 +1,11 @@
+package business.exceptions;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(value=HttpStatus.BAD_REQUEST)  // 
+public class InvalidLabNumber extends RuntimeException {
+    public InvalidLabNumber() {
+        super("Invalid lab number.");
+    }
+}
